# main/docs/swagger.json
{
    "swagger": "2.0",
    "info": {
        "description": "Simple Twitter-like REST API.",
        "title": "Twitter-like API",
        "contact": {},
        "version": "1.0"
    },
    "basePath": "/api/v1/",
    "paths": {
...
        "/protected/feed": {
            "get": {
                "security": [
                    {
                        "Bearer": []
                    }
                ],
                "description": "Get tweets feed for current authenticated user (subscriptions, own tweets, etc.).",
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "feed"
                ],
                "summary": "Get user feed",
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/response.TweetList"
                        }
                    },
                    "401": {
                        "description": "Unauthorized",
                        "schema": {
                            "$ref": "#/definitions/response.Error"
                        }
                    },
                    "500": {
                        "description": "Internal server error",
                        "schema": {
                            "$ref": "#/definitions/response.Error"
                        }
                    }
                }
            }
        },
...
        "/public/search": {
            "get": {
                "description": "Search users or tweets by query string. Use type=user to search users; otherwise searches tweets.",
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "search"
                ],
                "summary": "Search users or tweets",
                "parameters": [
                    {
                        "type": "string",
                        "description": "Search query",
                        "name": "query",
                        "in": "query",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/response.SearchResult"
                        }
                    },
                    "400": {
                        "description": "Missing or invalid query parameter",
                        "schema": {
                            "$ref": "#/definitions/response.Error"
                        }
                    },
                    "500": {
                        "description": "Search failed",
                        "schema": {
                            "$ref": "#/definitions/response.Error"
                        }
                    }
                }
            }
        },
...
    },
    "definitions": {
...
        "response.Counters": {
            "type": "object",
            "properties": {
                "like_count": {
                    "type": "integer"
                },
                "reply_count": {
                    "type": "integer"
                },
                "retweet_count": {
                    "type": "integer"
                }
            }
        },
        "response.Error": {
            "type": "object",
            "properties": {
                "error": {
                    "type": "string"
                }
            }
        },
...
        "response.SearchResult": {
            "type": "object",
            "properties": {
                "tweets": {
                    "type": "array",
                    "items": {
                        "$ref": "#/definitions/response.Tweet"
                    }
                },
                "users": {
                    "type": "array",
                    "items": {
                        "$ref": "#/definitions/response.User"
                    }
                }
            }
        },
...
        "response.SmallUser": {
            "type": "object",
            "properties": {
                "avatar_url": {
                    "type": "string"
                },
                "username": {
                    "type": "string"
                }
            }
        },
...
        "response.Tweet": {
            "type": "object",
            "properties": {
                "author": {
                    "$ref": "#/definitions/response.SmallUser"
                },
                "content": {
                    "type": "string"
                },
                "counters": {
                    "$ref": "#/definitions/response.Counters"
                },
                "created_at": {
                    "type": "string"
                },
                "id": {
                    "type": "integer"
                },
                "media_url": {
                    "type": "string"
                },
                "parent_tweet_id": {
                    "type": "integer"
                },
                "updated_at": {
                    "type": "string"
                }
            }
        },
        "response.TweetList": {
            "type": "object",
            "properties": {
                "tweets": {
                    "type": "array",
                    "items": {
                        "$ref": "#/definitions/response.Tweet"
                    }
                }
            }
        },
...
        "response.User": {
            "type": "object",
            "properties": {
                "avatar_url": {
                    "type": "string"
                },
                "bio": {
                    "type": "string"
                },
                "created_at": {
                    "type": "string"
                },
                "email": {
                    "type": "string"
                },
                "gen": {
                    "type": "string"
                },
                "id": {
                    "type": "integer"
                },
                "username": {
                    "type": "string"
                }
            }
        },
...
    },
    "securityDefinitions": {
        "Bearer": {
            "description": "Use \"Bearer {access_token}\" format.",
            "type": "apiKey",
            "name": "Authorization",
            "in": "header"
        }
    }
}